split demo magic into demoMagic folder
