+++ packages/demo/frontend/src/constants/borrow.ts
@@ -1,0 +1,2 @@
+/** Debt below half a cent is interest dust; treat the position as fully repaid. */
+export const DEBT_DUST_THRESHOLD = 0.005
