+++ packages/demo/frontend/src/demoMagic/aaveDemoMagic.ts
