+++ packages/demo/frontend/src/demoMagic/index.ts
@@ -1,0 +1,3 @@
+export * from './aaveDemoMagic'
+export * from './morphoDemoMagic'
+export { ReacquireDebtNotice } from './ReacquireDebtNotice'
